Bugfix: error when GPKG does not exists before saving offline
When the destination GeoPackage does not exists before converting
project for offline editing, QGIS is not able to save the project, that
is: to create an empty GeoPackage and save the project inside it.

So the new workflow is the following: save the project in a temporary
QGZ file (this step is needed so that the project can be set to save
layer paths *relative* to the project home folder), then convert the
project offline in a GeoPackage in the same folder than the QGZ file
(and since paths are relative, we get a bunch of `./` paths), then save
the project again in the same GeoPackage than the layers, and then
delete the QGZ file.